Parameter type fixed

+++ portable/comp-transfer/src/main/java/org/bosik/diacomp/core/services/preferences/PreferencesTypedService.java
@@ -188,7 +188,7 @@ public class PreferencesTypedService implements PreferencesService
 	 * @param preferenceID
 	 * @param value
 	 */
-	public void setDoubleValue(PreferenceID preferenceID, Float value)
+	public void setDoubleValue(PreferenceID preferenceID, Double value)
 	{
 		service.setString(buildEntry(preferenceID, String.valueOf(value)));
 	}
